vdiff: factor Summarize out from vtctldclient
Factor out buildSingleSummary as vdiff.Summarize, which generates a summary of
the VDiff operation based on the VDiffShow RPC response.

Fixes #16780.
