+++ go/cmd/vtctldclient/command/vreplication/vdiff/vdiff_test.go
@@ -19,7 +19,6 @@ package vdiff
 import (
 	"context"
 	"fmt"
-	"math"
 	"testing"
 	"time"
 
@@ -689,113 +688,4 @@ func TestGetStructNames(t *testing.T) {
 	got := getStructFieldNames(s{})
 	want := []string{"A", "B"}
 	require.EqualValues(t, want, got)
-}
-
-func TestBuildProgressReport(t *testing.T) {
-	now := time.Now()
-	type args struct {
-		summary       *summary
-		rowsToCompare int64
-	}
-	tests := []struct {
-		name string
-		args args
-		want *vdiff.ProgressReport
-	}{
-		{
-			name: "no progress",
-			args: args{
-				summary:       &summary{RowsCompared: 0},
-				rowsToCompare: 100,
-			},
-			want: &vdiff.ProgressReport{
-				Percentage: 0,
-				ETA:        "", // no ETA
-			},
-		},
-		{
-			name: "one third of the way",
-			args: args{
-				summary: &summary{
-					RowsCompared: 33,
-					StartedAt:    now.Add(-10 * time.Second).UTC().Format(vdiff.TimestampFormat),
-				},
-				rowsToCompare: 100,
-			},
-			want: &vdiff.ProgressReport{
-				Percentage: 33,
-				ETA:        now.Add(20 * time.Second).UTC().Format(vdiff.TimestampFormat),
-			},
-		},
-		{
-			name: "half way",
-			args: args{
-				summary: &summary{
-					RowsCompared: 5000000000,
-					StartedAt:    now.Add(-10 * time.Hour).UTC().Format(vdiff.TimestampFormat),
-				},
-				rowsToCompare: 10000000000,
-			},
-			want: &vdiff.ProgressReport{
-				Percentage: 50,
-				ETA:        now.Add(10 * time.Hour).UTC().Format(vdiff.TimestampFormat),
-			},
-		},
-		{
-			name: "full progress",
-			args: args{
-				summary: &summary{
-					RowsCompared: 100,
-					CompletedAt:  now.UTC().Format(vdiff.TimestampFormat),
-				},
-				rowsToCompare: 100,
-			},
-			want: &vdiff.ProgressReport{
-				Percentage: 100,
-				ETA:        now.UTC().Format(vdiff.TimestampFormat),
-			},
-		},
-		{
-			name: "more than in I_S",
-			args: args{
-				summary: &summary{
-					RowsCompared: 100,
-					CompletedAt:  now.UTC().Format(vdiff.TimestampFormat),
-				},
-				rowsToCompare: 50,
-			},
-			want: &vdiff.ProgressReport{
-				Percentage: 100,
-				ETA:        now.UTC().Format(vdiff.TimestampFormat),
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			buildProgressReport(tt.args.summary, tt.args.rowsToCompare)
-			// We always check the percentage
-			require.Equal(t, int(tt.want.Percentage), int(tt.args.summary.Progress.Percentage))
-
-			// We only check the ETA if there is one.
-			if tt.want.ETA != "" {
-				// Let's check that we're within 1 second to avoid flakes.
-				wantTime, err := time.Parse(vdiff.TimestampFormat, tt.want.ETA)
-				require.NoError(t, err)
-				var timeDiff float64
-				if tt.want.Percentage == 100 {
-					completedTime, err := time.Parse(vdiff.TimestampFormat, tt.args.summary.CompletedAt)
-					require.NoError(t, err)
-					timeDiff = math.Abs(completedTime.Sub(wantTime).Seconds())
-				} else {
-					startTime, err := time.Parse(vdiff.TimestampFormat, tt.args.summary.StartedAt)
-					require.NoError(t, err)
-					completedTimeUnix := float64(now.UTC().Unix()-startTime.UTC().Unix()) * (100 / tt.want.Percentage)
-					estimatedTime, err := time.Parse(vdiff.TimestampFormat, tt.want.ETA)
-					require.NoError(t, err)
-					timeDiff = math.Abs(estimatedTime.Sub(startTime).Seconds() - completedTimeUnix)
-				}
-				require.LessOrEqual(t, timeDiff, 1.0)
-			}
-		})
-	}
 }
