+++ go/vt/vttablet/tabletmanager/vdiff/summary_test.go
@@ -1,0 +1,134 @@
+/*
+Copyright 2024 The Vitess Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package vdiff
+
+import (
+	"math"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestBuildProgressReport(t *testing.T) {
+	now := time.Now()
+	type args struct {
+		summary       *Summary
+		rowsToCompare int64
+	}
+	tests := []struct {
+		name string
+		args args
+		want *ProgressReport
+	}{
+		{
+			name: "no progress",
+			args: args{
+				summary:       &Summary{RowsCompared: 0},
+				rowsToCompare: 100,
+			},
+			want: &ProgressReport{
+				Percentage: 0,
+				ETA:        "", // no ETA
+			},
+		},
+		{
+			name: "one third of the way",
+			args: args{
+				summary: &Summary{
+					RowsCompared: 33,
+					StartedAt:    now.Add(-10 * time.Second).UTC().Format(TimestampFormat),
+				},
+				rowsToCompare: 100,
+			},
+			want: &ProgressReport{
+				Percentage: 33,
+				ETA:        now.Add(20 * time.Second).UTC().Format(TimestampFormat),
+			},
+		},
+		{
+			name: "half way",
+			args: args{
+				summary: &Summary{
+					RowsCompared: 5000000000,
+					StartedAt:    now.Add(-10 * time.Hour).UTC().Format(TimestampFormat),
+				},
+				rowsToCompare: 10000000000,
+			},
+			want: &ProgressReport{
+				Percentage: 50,
+				ETA:        now.Add(10 * time.Hour).UTC().Format(TimestampFormat),
+			},
+		},
+		{
+			name: "full progress",
+			args: args{
+				summary: &Summary{
+					RowsCompared: 100,
+					CompletedAt:  now.UTC().Format(TimestampFormat),
+				},
+				rowsToCompare: 100,
+			},
+			want: &ProgressReport{
+				Percentage: 100,
+				ETA:        now.UTC().Format(TimestampFormat),
+			},
+		},
+		{
+			name: "more than in I_S",
+			args: args{
+				summary: &Summary{
+					RowsCompared: 100,
+					CompletedAt:  now.UTC().Format(TimestampFormat),
+				},
+				rowsToCompare: 50,
+			},
+			want: &ProgressReport{
+				Percentage: 100,
+				ETA:        now.UTC().Format(TimestampFormat),
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			buildProgressReport(tt.args.summary, tt.args.rowsToCompare)
+			// We always check the percentage
+			require.Equal(t, int(tt.want.Percentage), int(tt.args.summary.Progress.Percentage))
+
+			// We only check the ETA if there is one.
+			if tt.want.ETA != "" {
+				// Let's check that we're within 1 second to avoid flakes.
+				wantTime, err := time.Parse(TimestampFormat, tt.want.ETA)
+				require.NoError(t, err)
+				var timeDiff float64
+				if tt.want.Percentage == 100 {
+					completedTime, err := time.Parse(TimestampFormat, tt.args.summary.CompletedAt)
+					require.NoError(t, err)
+					timeDiff = math.Abs(completedTime.Sub(wantTime).Seconds())
+				} else {
+					startTime, err := time.Parse(TimestampFormat, tt.args.summary.StartedAt)
+					require.NoError(t, err)
+					completedTimeUnix := float64(now.UTC().Unix()-startTime.UTC().Unix()) * (100 / tt.want.Percentage)
+					estimatedTime, err := time.Parse(TimestampFormat, tt.want.ETA)
+					require.NoError(t, err)
+					timeDiff = math.Abs(estimatedTime.Sub(startTime).Seconds() - completedTimeUnix)
+				}
+				require.LessOrEqual(t, timeDiff, 1.0)
+			}
+		})
+	}
+}
